backend/requirements.txt
openai
PyMuPDF
tiktoken
python-dotenv
sqlalchemy
pgvector
psycopg2
tqdm
fastapi
uvicorn
python-multipart